Add array replace functions
