+++ datafusion/functions-array/Cargo.toml
@@ -40,6 +40,7 @@ path = "src/lib.rs"
 arrow = { workspace = true }
 arrow-array = { workspace = true }
 arrow-buffer = { workspace = true }
+arrow-ord = { workspace = true }
 arrow-schema = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-execution = { workspace = true }
