+++ datafusion/physical-expr/src/lib.rs
@@ -17,7 +17,6 @@
 
 pub mod aggregate;
 pub mod analysis;
-pub mod array_expressions;
 pub mod binary_map;
 pub mod conditional_expressions;
 pub mod equivalence;
